refactor(resilientws): split into modular files and fix concurrent write races
